feat(android): migrate control/websocket protocol to generated Kotlin SDK
Ticket: T2.5

ControlSession/ControlViewModel use com.stream.v1.Envelope and
generated payload types via JsonFormat, replacing the deleted
domain.model.control.Envelope and kotlinx.serialization DTOs.

Adds EnvelopeProto.kt (buildEnvelope/decodePayload), mirroring
macOS's Envelope+Proto.swift bridge pattern.

CastSheet.kt/RemoteScreen.kt updated for the new SessionInfo import
and double/Float volume conversion at UI boundaries.

Closes T2.5 — Android fully migrated (REST + control protocol,
controller-only scope per the design doc).

+++ apps/android/app/src/main/java/com/nativestream/android/ui/screens/remote/RemoteScreen.kt
@@ -62,9 +62,9 @@ fun RemoteScreen(
     val isPullingBack by controlViewModel.isPullingBack.collectAsState()
     val session       = sessions.firstOrNull { it.playing } ?: return
 
-    var volume by remember(session.deviceId) { mutableFloatStateOf(session.volume) }
+    var volume by remember(session.deviceId) { mutableFloatStateOf(session.volume.toFloat()) }
 
-    LaunchedEffect(session.volume) { volume = session.volume }
+    LaunchedEffect(session.volume) { volume = session.volume.toFloat() }
 
     LaunchedEffect(Unit) {
         controlViewModel.pullBackReady.collect { ready ->
